% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier package.R
\name{outlier}
\alias{outlier}
\title{Outlier detection}
\usage{
outlier(data)
}
\arguments{
\item{data=}{Dataframe}
}
\value{
It will return DataFrame having outlier count on original data and after log transformation
}
\description{
It will provide Outlier count with original data and after doing log transformation in the form of DataFrame
}
\author{
Tejas Vartak, Shivani Goyal
}
